+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 # Next Release
 
 - [#223](https://github.com/iiasa/message_ix/pull/223): Explicit MACRO parameterization and calibration support
+- [#255](https://github.com/iiasa/message_ix/pull/249): Add :mod:`message_ix.testing.nightly` and `nightly` CLI command group for slow-running tests.
+- [#249](https://github.com/iiasa/message_ix/pull/249): Build MESSAGE and MESSAGE_MACRO classes on ixmp model API; adjust Scenario.
 - [#235](https://github.com/iiasa/message_ix/pull/236): Add a reporting tutorial.
 - [#236](https://github.com/iiasa/message_ix/pull/236): Enhance reporting.
 - [#232](https://github.com/iiasa/message_ix/pull/232): Add Westeros tutorial for modelling seasonality, update existing tutorials.